Split out the existing user views into modules
