+++ timetable_solutions/interfaces/users/views/__init__.py
@@ -1,0 +1,9 @@
+from ._authentication import Login, custom_logout
+from ._dashboard import Dashboard
+from ._password_management import PasswordChange
+from ._register import (
+    ProfileRegistration,
+    Register,
+    SchoolRegisterPivot,
+    SchoolRegistration,
+)
