+++ timetable_solutions/interfaces/users/views/_register.py
@@ -1,5 +1,5 @@
 """
-Views relating to user authentication and registration.
+Views for user registration.
 
 User registration has the following steps:
 Step 1 - provide basic details (name, email address, password etc.)
@@ -12,9 +12,7 @@ Step 3b - the user must provide a school access key to associate themselves with
 # Django imports
 from django import http
 from django.contrib import messages
-from django.contrib.auth import forms as auth_forms
-from django.contrib.auth import login, logout, mixins
-from django.contrib.auth import views as auth_views
+from django.contrib.auth import login, logout
 from django.db import transaction
 from django.shortcuts import redirect
 from django.urls import reverse
@@ -128,57 +126,3 @@ class ProfileRegistration(generic.FormView):
         )
         messages.success(self.request, message=message)
         return super().form_valid(form=form)
-
-
-class Login(auth_views.LoginView):
-    """
-    Add some extra rules to the login process.
-    """
-
-    def setup(self, request: http.HttpRequest, *args: object, **kwargs: object) -> None:
-        """
-        Log a user out if they visit the login page.
-        """
-        super().setup(request, *args, **kwargs)
-        if request.user.is_authenticated:
-            logout(request)
-
-    def form_valid(self, form: auth_forms.AuthenticationForm) -> http.HttpResponse:
-        user = form.get_user()
-        if not hasattr(user, "profile"):
-            # The user has not completed registration
-            login(
-                self.request, user, backend="django.contrib.auth.backends.ModelBackend"
-            )
-            return redirect(to=UrlName.REGISTER_PIVOT.url())
-        elif user.profile.approved_by_school_admin:
-            return super().form_valid(form)
-        else:
-            error_message = (
-                "Your account has not yet been approved by your school's admin account.\n"
-                "Please contact them directly to approve your account."
-            )
-            form.add_error(None, error_message)
-            return super().form_invalid(form=form)
-
-
-def custom_logout(request: http.HttpResponse) -> http.HttpResponseRedirect:
-    """
-    Redirect users to the login page when they log out.
-    """
-    if request.user.is_authenticated:
-        logout(request)
-    return redirect(UrlName.LOGIN.url())
-
-
-class PasswordChange(auth_views.PasswordChangeView):
-    success_url = UrlName.DASHBOARD.url(lazy=True)
-
-    def form_valid(self, form: auth_forms.PasswordChangeForm) -> http.HttpResponse:
-        message = "Your password has been updated!"
-        messages.success(request=self.request, message=message)
-        return super().form_valid(form=form)
-
-
-class Dashboard(mixins.LoginRequiredMixin, generic.TemplateView):
-    template_name = "users/dashboard.html"
